agentic-sql-ai/backend/requirements.txt
fastapi
uvicorn
openai
sqlparse
pydantic
pyyaml
snowflake-connector-python